dnsbl-milter.c: Fixed bug where no msgid is returned
When the milter assumes that smfi_getsymval() would return /something/,
it crashes in a spectacular fashion:  it completely locks up my web
server for about 30 seconds before it gets around to finally dumping
core out from under me.

The documentation for libmilter is fairly sketchy, but I believe
that it says that milters can't assume that any sendmail variables
will actually be available.    This patch checks for the condition
that no variable is there, and if it isn't it sets msgid to a dummy
value.